@@ -1,24 +1,32 @@
 package com.syntax.note.login;
 
 import android.content.Intent;
-import android.support.design.widget.Snackbar;
-import android.support.design.widget.TextInputLayout;
-import android.support.v7.app.AppCompatActivity;
+
+import com.google.android.gms.auth.api.signin.GoogleSignIn;
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
+import com.google.android.gms.auth.api.signin.GoogleSignInClient;
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
+import com.google.android.gms.common.api.ApiException;
+import com.google.android.gms.tasks.Task;
+import com.google.android.material.snackbar.Snackbar;
+import com.google.android.material.textfield.TextInputLayout;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.ImageButton;
 import android.widget.ProgressBar;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.syntax.note.R;
 import com.syntax.note.home.HomeActivity;
 import com.syntax.note.signinRequestPOJO.signinRequestBean;
 import com.syntax.note.signinRequestPOJO.Data;
 import com.syntax.note.signinResponsePOJO.signinResponseBean;
+import com.syntax.note.socialRequestPOJO.socialRequestBean;
 import com.syntax.note.utility.Constant;
 import com.syntax.note.utility.SharePreferenceUtils;
 import com.syntax.note.webServices.ServiceInterface;
@@ -30,6 +38,7 @@ import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
 
 public class SigninActivity extends AppCompatActivity {
+    private static final int RC_SIGN_IN = 121;
     EditText syntaxSigninEmail,syntaxSigninPassword;
     Button syntaxSigninBtn;
     String syntax_email,syntax_password;
@@ -43,6 +52,10 @@ public class SigninActivity extends AppCompatActivity {
     boolean isValid=false;
     private ProgressBar pBar;
 
+    GoogleSignInClient mGoogleSignInClient;
+
+    ImageButton google;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -52,6 +65,18 @@ public class SigninActivity extends AppCompatActivity {
         setUpWidget();
         pBar.setVisibility(View.GONE);
         getData();
+
+
+        // Configure sign-in to request the user's ID, email address, and basic
+// profile. ID and basic profile are included in DEFAULT_SIGN_IN.
+        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                .requestEmail()
+                .build();
+
+        // Build a GoogleSignInClient with the options specified by gso.
+        mGoogleSignInClient = GoogleSignIn.getClient(this, gso);
+
+
 
         //Retrofit
 
@@ -95,7 +120,105 @@ public class SigninActivity extends AppCompatActivity {
             }
         });
 
+        google.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+
+                Intent signInIntent = mGoogleSignInClient.getSignInIntent();
+                startActivityForResult(signInIntent, RC_SIGN_IN);
+
+            }
+        });
+
     }
+
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        // Result returned from launching the Intent from GoogleSignInClient.getSignInIntent(...);
+        if (requestCode == RC_SIGN_IN) {
+            // The Task returned from this call is always completed, no need to attach
+            // a listener.
+            Task<GoogleSignInAccount> task = GoogleSignIn.getSignedInAccountFromIntent(data);
+            handleSignInResult(task);
+        }
+    }
+
+
+    private void handleSignInResult(Task<GoogleSignInAccount> completedTask) {
+        try {
+            GoogleSignInAccount account = completedTask.getResult(ApiException.class);
+
+            // Signed in successfully, show authenticated UI.
+
+            String email = account.getEmail();
+            String pid = account.getId();
+
+
+            Log.d("email" , email);
+            Log.d("pid" , pid);
+
+            pBar.setVisibility(View.VISIBLE);
+
+            socialRequestBean body = new socialRequestBean();
+            body.setAction("social_login");
+            com.syntax.note.socialRequestPOJO.Data data = new com.syntax.note.socialRequestPOJO.Data();
+
+            data.setEmail(email);
+            data.setPid(pid);
+            body.setData(data);
+
+
+            Call<signinResponseBean> call = serviceInterface.socialsignin(body);
+
+
+            call.enqueue(new Callback<signinResponseBean>() {
+                @Override
+                public void onResponse(Call<signinResponseBean> call, Response<signinResponseBean> response) {
+
+                    if (response.body().getStatus().equals("1"))
+                    {
+                        pBar.setVisibility(View.GONE);
+                        // Toast.makeText(SigninActivity.this, ""+response.body()
+                        //  .getData().getName(), Toast.LENGTH_SHORT).show();
+                        Log.i("signin",response.body().getData().getEmail());
+                        SharePreferenceUtils.getInstance().saveString(Constant.USER_id,response.body().getData().getUserId());
+                        SharePreferenceUtils.getInstance().saveString(Constant.USER_email,response.body().getData().getEmail());
+                        SharePreferenceUtils.getInstance().saveString(Constant.USER_name,response.body().getData().getName());
+                        SharePreferenceUtils.getInstance().saveString(Constant.USER_phone,response.body().getData().getMobile());
+                        Intent homeIntent = new Intent(SigninActivity.this,HomeActivity.class);
+                        startActivity(homeIntent);
+                        finish();
+                    }
+                    else
+                    {
+                        pBar.setVisibility(View.GONE);
+                        Snackbar snackbar = Snackbar.make(rootlayout,"Login Failed Check Login Credential",Snackbar.LENGTH_LONG);
+                        View snackBarView = snackbar.getView();
+                        snackBarView.setBackgroundColor(getResources().getColor(R.color.seaGreen));
+                        snackbar.show();
+
+                    }
+
+                }
+
+                @Override
+                public void onFailure(Call<signinResponseBean> call, Throwable t) {
+                    pBar.setVisibility(View.GONE);
+                }
+            });
+
+
+        } catch (ApiException e) {
+            // The ApiException status code indicates the detailed failure reason.
+            // Please refer to the GoogleSignInStatusCodes class reference for more information.
+            Log.w("asdad", "signInResult:failed code=" + e.getStatusCode());
+
+        }
+    }
+
 
     private void signin() {
         signinRequestBean body = new signinRequestBean();
@@ -156,8 +279,9 @@ public class SigninActivity extends AppCompatActivity {
         rootlayout = findViewById(R.id.rootlayout);
         inputLayoutSignin = findViewById(R.id.inputLayoutEmail);
         inputLayoutPassword = findViewById(R.id.inputLayoutPassword);
+        google = findViewById(R.id.google);
 
-        pBar = (ProgressBar)findViewById(R.id.progressBar);
+        pBar = findViewById(R.id.progressBar);
     }
 
     private void getData() {
